pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartwatch-automator"
version = "0.1.0"
description = "A configurable file watcher and automator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your@email.com" }
]
keywords = ["automation", "file-watcher", "watchdog", "cli"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
]
dependencies = [
    "watchdog>=4.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
]

[project.scripts]
smartwatch = "smartwatch.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/smartwatch-automator"
Issues = "https://github.com/yourusername/smartwatch-automator/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["smartwatch"]